Add footer style and other minor adjustments
In order to fix a problem with the webpage not stretching correctly I
had to change the height setting of the entire body. Even now I am not
sure that it was the right solution but in the end the webpage ended
up working the way I wanted it to.